Refactor: Relocate protocol definitions to config.py to resolve import cycles (#39)
* Initial plan

* Relocate protocol definitions from person.py and elevator.py to config.py to resolve import cycles


* Add flake8 suppression for protocol definitions in config.py


---------
