+++ mytower/game/core/config.py
@@ -2,19 +2,111 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Final
+from typing import TYPE_CHECKING, Final, Protocol
 
 from mytower.game.core.units import Blocks, Meters, Time, Velocity
+from mytower.game.core.types import RGB
 
 if TYPE_CHECKING:
-    from mytower.game.core.types import RGB
-    from mytower.game.entities.elevator import ElevatorConfigProtocol, ElevatorCosmeticsProtocol
     from mytower.game.views.desktop_ui import UIConfigProtocol
-    from mytower.game.entities.person import PersonConfigProtocol, PersonCosmeticsProtocol
 
 # It can't seem to make up its mind whether to use snake_case because they are class properties or UPPER_CASE for constants
 # So, it gets neither
-# pylint: disable=invalid-name 
+# pylint: disable=invalid-name
+
+# flake8: noqa: E704
+# Protocol definitions for configuration
+class ElevatorConfigProtocol(Protocol):
+    """Config requirements for Elevator class"""
+
+    @property
+    def MAX_SPEED(self) -> Velocity: ...
+
+    @property
+    def MAX_CAPACITY(self) -> int: ...
+
+    @property
+    def PASSENGER_LOADING_TIME(self) -> Time: ...
+
+    @property
+    def IDLE_LOG_TIMEOUT(self) -> Time: ...
+
+    @property
+    def MOVING_LOG_TIMEOUT(self) -> Time: ...
+
+    @property
+    def IDLE_WAIT_TIMEOUT(self) -> Time: ...
+
+class ElevatorCosmeticsProtocol(Protocol):
+    """Visual appearance settings for Elevator class"""
+
+    @property
+    def SHAFT_COLOR(self) -> RGB: ...
+
+    @property
+    def SHAFT_OVERHEAD_COLOR(self) -> RGB: ...
+
+    @property
+    def CLOSED_COLOR(self) -> RGB: ...
+
+    @property
+    def OPEN_COLOR(self) -> RGB: ...
+
+    @property
+    def SHAFT_OVERHEAD_HEIGHT(self) -> Meters: ...
+    
+    @property
+    def ELEVATOR_WIDTH(self) -> Meters: ...
+    
+    @property
+    def ELEVATOR_HEIGHT(self) -> Meters: ...
+
+class PersonConfigProtocol(Protocol):
+    """Config requirements for Person class"""
+
+    @property
+    def MAX_SPEED(self) -> Velocity: ...
+
+    @property
+    def MAX_WAIT_TIME(self) -> Time: ...
+
+    @property
+    def IDLE_TIMEOUT(self) -> Time: ...
+
+    @property
+    def RADIUS(self) -> Meters: ...
+
+class PersonCosmeticsProtocol(Protocol):
+    """Visual appearance settings for Person class"""
+
+    @property
+    def ANGRY_MAX_RED(self) -> int: ...
+
+    @property
+    def ANGRY_MIN_GREEN(self) -> int: ...
+
+    @property
+    def ANGRY_MIN_BLUE(self) -> int: ...
+
+    @property
+    def INITIAL_MAX_RED(self) -> int: ...
+
+    @property
+    def INITIAL_MAX_GREEN(self) -> int: ...
+
+    @property
+    def INITIAL_MAX_BLUE(self) -> int: ...
+
+    @property
+    def INITIAL_MIN_RED(self) -> int: ...
+
+    @property
+    def INITIAL_MIN_GREEN(self) -> int: ...
+
+    @property
+    def INITIAL_MIN_BLUE(self) -> int: ...
+
+# Concrete configuration implementations
 @dataclass
 class ElevatorConfig:
     """Concrete implementation of Elevator configuration"""
